+++ RELEASE_NOTES.rst
@@ -5,7 +5,8 @@ Release notes for bcolz
 Changes from 0.8.0 to 0.8.1
 ===========================
 
-- pass
+- Implement ``nleftover`` which returns the number of leftover and uncompressed
+  elements.
 
 Changes from 0.7.3 to 0.8.0
 ===========================